+++ packages/xo-server-backup-reports/src/index.js
@@ -107,6 +107,10 @@ class BackupReportsXoPlugin {
   constructor (xo) {
     this._xo = xo
     this._report = this._wrapper.bind(this)
+    this._xo.addApiMethod(
+      'plugin.backupReport.send',
+      this._backupNgListener.bind(this)
+    )
   }
 
   configure ({ toMails, toXmpp }) {
@@ -126,13 +130,13 @@ class BackupReportsXoPlugin {
     return new Promise(resolve =>
       resolve(
         job.type === 'backup'
-          ? this._backupNgListener(status, job, schedule, runJobId)
+          ? this._backupNgListener({ schedule, runJobId })
           : this._listener(status, job, schedule, runJobId)
       )
     ).catch(logError)
   }
 
-  async _backupNgListener (_1, _2, { timezone }, runJobId) {
+  async _backupNgListener ({ schedule, runJobId }) {
     const xo = this._xo
     const log = await xo.getBackupNgLogs(runJobId)
 
@@ -144,8 +148,12 @@ class BackupReportsXoPlugin {
       return
     }
 
+    if (schedule === undefined) {
+      schedule = await xo.getSchedule(log.scheduleId)
+    }
+
     const jobName = (await xo.getJob(log.jobId, 'backup')).name
-    const formatDate = createDateFormater(timezone)
+    const formatDate = createDateFormater(schedule.timezone)
     const getTemporalDataMarkdown = createGetTemporalDataMarkdown(formatDate)
 
     if (